requirements.txt
kafka-python==2.0.2
confluent-kafka==2.3.0
faker==19.12.0
python-dotenv==1.0.0